Added methods to Colours class to access colour codes directly
It would be handy to expose the colour code directly to increase the
flexibility of the class (with potential efficiency gains in output
size where many lines are coloured). Three methods were added to
enable this:

 - getForegroundCode($foregroundColour)
 - getBackgroundCode($backgroundColour)
 - getResetCode()

In addition, the string() method was adapted to make use of these.

Added unit tests for new Colours class methods

Simple unit tests for:
 - getForegroundCode($foregroundColour)
 - getBackgroundCode($backgroundColour)
 - getResetCode()